+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [2.7.0] - 2025-03-31
+
 - Add new `dateString` decoder (see [docs](https://decoders.cc/api.html#dateString))
 - Add new `.refineType<SubT>()` method (see
   [docs](https://decoders.cc/Decoder.html#refineType))